v8

+++ Controlador/Ctrl_Administrador/MySQL_Borrar_Suministro.php
@@ -18,45 +18,52 @@ if (isset($_SESSION['usuario'])) {
         // Crear una instancia de la clase Database
         $conn = new Conexion_BD();
 
-        // Construir la consulta SQL para eliminar en Primera instancia los producto de la tabla requerimientos.
-        $sql = "DELETE FROM requerimientos WHERE id_producto = ?";
+        // Construir la consulta SQL para obtener el producto
+        $sql = "SELECT producto FROM producto WHERE id_producto = ?";
         $stmt = $conn->prepareStatement($sql);
         $stmt->bind_param("i", $idSuministro);
         $stmt->execute();
+        $result = $stmt->get_result();
 
-        // Construir la consulta SQL para eliminar en Segunda instancia el Suministro.
-        $sql = "DELETE FROM producto WHERE id_producto = ?";
-        $stmt = $conn->prepareStatement($sql);
-        $stmt->bind_param("i", $idSuministro);
-        $stmt->execute();
-
-        // Verificar si la eliminación afectó a alguna fila
-        if ($stmt->affected_rows > 0) {
-
-            // La eliminación fue exitosa
-            $_SESSION['idSuministro'] = $idSuministro;
-            header("Location: ../../Vista/Administrador/Adm_Gestion_Suministros.php");            
+        // Verificar si se encontró el producto
+        if ($result->num_rows > 0) {
             
+            $productoCon = $result->fetch_assoc();
+            $_SESSION['Producto'] = $productoCon['nombre']; // Guarda el nombre del producto en la sesión
+
+            // Construir la consulta SQL para eliminar en Primera instancia los productos de la tabla requerimientos.
+            $sql = "DELETE FROM requerimientos WHERE id_producto = ?";
+            $stmt = $conn->prepareStatement($sql);
+            $stmt->bind_param("i", $idSuministro);
+            $stmt->execute();
+
+            // Construir la consulta SQL para eliminar en Segunda instancia el Suministro.
+            $sql = "DELETE FROM producto WHERE id_producto = ?";
+            $stmt = $conn->prepareStatement($sql);
+            $stmt->bind_param("i", $idSuministro);
+            $stmt->execute();
+
+            // Verificar si la eliminación afectó a alguna fila
+            if ($stmt->affected_rows > 0) {
+                // La eliminación fue exitosa
+                header("Location: ../../Vista/Administrador/Adm_Gestion_Suministros.php");
+            } else {
+                echo("La eliminación no afectó a ninguna fila (posiblemente el ID SUMINISTRO no existe)");
+            }
         } else {
-
-            echo("La eliminación no afectó a ninguna fila (posiblemente el ID SUMINISTRO no existe)");
-
+            echo("No se encontró ningún producto con el ID proporcionado");
         }
-        
-    } else {
 
+    } else {
         // Redireccionar a la página de gestión de Suministros
         header("Location: ../../Vista/Administrador/Adm_Gestion_Suminitros.php");
         exit();
-
     }
 
 } else {
-
     header("Location: ../../Controlador/Utilidades/Sesion_Destroy.php");
     exit();
-
-}    
+}
 
 // Cerrar la conexión a la base de datos
 $conn->closeConnection();
